expose spec loading from wasm onto Spec type
